fix for launching outside of project dir
If the process was launched outside of the project dir, git-rev would not properly identify the project git id because it would default to CWD (which is not current). This changes git-rev to use the `__dirname` of the file which is more likely to reside inside of the repo the user wants to id.