ability to change current workig directory
This module offers a nice way to get current hash of an npm project
that is currently in the git repo's root.

Some hosting providers run and serve npm projects from the repo root,
but in MS's Azure, the git `repository` and the `wwwroot` directories
are sibling repos.  During deployment phase, the files are deployed to
`wwwroot`.  This change makes it possible for an npm project running
in Azure (or any provider where the npm project is run or deployed in
a different location from the git repo root) to re-use this module to
easily grab the last hash.