Use the git packfile if /refs/heads fails
When getting the current commit, looking in /refs/heads can fail. This
is because git sometimes stores these refs in a packfile for performance
reasons
(https://www.kernel.org/pub/software/scm/git/docs/git-pack-refs.html).
If we can't find the commit in /refs/heads, we should fall back to
look in the packfile.